Integration of CPixmap.
QPixmap is now replaced by CPixmap. This class allow some effects on the picture,
like NVG, black and white, and so on.
Now, it must be improved, to add a possibility of reverting the effects ( for instance,
if you apply black and white effect, you should be able to get a colored picture again ).