Translate help text, and improved BrowserItem
BrowserItem didn't take care of the user language, so it was impossible to translate the introduction text.
I nearly fixed that, but there is still a little problem : I don't know how to test if a qrc exists ! (so for now I tested if the file was there :S )
I translated the text.