Redo on delete works!
When deleting, replace the pointer towards the now unexisting adress of the deleted content by the new adress. That way, no segfault.
The same thing has to be done for the other commands (new content, text, and so on...).
It can probably be quite nicely generalised in AbstractCommand. That way, there would just be to call replaceContent where needed, and be done with it :)