[undo] Hide content instead of deleting it
It's the 21st century, we can afford to keep a few dozens of
images/text/whatever in the command stack. Thus I have changed my mind
and instead of doing the unnecessarely complicated bookkeeping of
actually deleting content from memory, and reloading it from a saved
on-disk copy, then updating the command stack (insane right?), we simply
hide the element upon deletion, and show it back when undoing the
deletion.

This has additional advantages, such as keeping the undo/redo stack of
sub-canvas intact upon restoring them.